update collateral WETH manager
